pyproject.toml
[project]
name = "homelab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "linkchecker>=10.5.0",
    "python-slugify>=8.0.4",
]
